+++ app/src/main/java/com/thingsflow/internapplication/AppModuleCoroutine.kt
@@ -4,7 +4,6 @@ import android.content.Context
 import androidx.room.Room
 import com.thingsflow.internapplication.data.IssueDao
 import com.thingsflow.internapplication.data.IssueDatabase
-import com.thingsflow.internapplication.data.RepositoryDao
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
@@ -48,10 +47,5 @@ object AppModuleCoroutine {
     @Provides
     fun provideIssueDao(issueDatabase: IssueDatabase): IssueDao{
         return issueDatabase.issueDao()
-    }
-
-    @Provides
-    fun provideRepositoryDao(issueDatabase: IssueDatabase): RepositoryDao{
-        return issueDatabase.repositoryDao()
     }
 }
