+++ app/src/main/java/com/thingsflow/internapplication/data/IssueDao.kt
@@ -12,14 +12,3 @@ interface IssueDao {
     @Insert(onConflict = REPLACE)
     suspend fun insertAllIssues(repositoryWithIssue: RepositoryWithIssue)
 }
-
-@Dao
-interface RepositoryDao {
-
-    @Query("SELECT * FROM repositories WHERE organization = :organization and repository = :repository")
-    suspend fun getRepository(organization: String, repository: String): RepositoryInfo
-
-    @Insert(onConflict = REPLACE)
-    suspend fun insertRepository(repositoryInfo: RepositoryInfo)
-
-}
