+++ app/src/main/java/com/thingsflow/internapplication/data/IssueDatabase.kt
@@ -6,12 +6,10 @@ import androidx.room.TypeConverters
 
 @Database(
     entities = [
-        RepositoryInfo::class,
         RepositoryWithIssue::class
     ], version = 1
 )
 @TypeConverters(IssueTypeConverter::class)
 abstract class IssueDatabase : RoomDatabase() {
     abstract fun issueDao(): IssueDao
-    abstract fun repositoryDao(): RepositoryDao
 }
