+++ app/src/main/java/com/thingsflow/internapplication/data/Item.kt
@@ -27,18 +27,9 @@ data class UserInfo(
     val userProfile: String
 )
 
-@Entity(tableName = "repositories", primaryKeys = ["organization", "repository"])
 data class RepositoryInfo(
     val organization: String,
     val repository: String
-)
-
-@Entity(tableName = "repository_with_issue", primaryKeys = ["organization", "repository"])
-data class RepositoryWithIssue(
-    val organization: String,
-    val repository: String,
-    @ColumnInfo(name = "issue_list")
-    val issueList: ArrayList<Item.IssueData>
 )
